Reduce worker nodes number, disable sysctls

+++ talos/init.sh
@@ -11,16 +11,14 @@ main()
     printf 'Generating Talos configuration...\n'
     rm -rf "$TALOS_CONFIG_HOME"
     talosctl gen config "$CLUSTER_NAME" "https://$CONTROL_PLANE_IP:6443" --output-dir "$TALOS_CONFIG_HOME" --install-image "$FACTORY_IMAGE" \
-        --config-patch @patch/common/sysctl.yaml \
-        --config-patch @patch/common/gatewayapi.yaml \
-        --config-patch @patch/common/cni-proxy.yaml \
         --config-patch @patch/common/diskSelector.yaml \
-        --config-patch @patch/common/longhorn.yaml
+        --config-patch @patch/common/longhorn.yaml \
+        --config-patch @patch/common/cni-proxy.yaml
 
     ln -sf "$CLUSTER_NAME/talosconfig" "$XDG_CONFIG_HOME/talos/config.yaml"
 
     wait 'Press enter to apply configuration to control plane nodes...'
-    sed 1d nodes.csv | grep controlplane | while IFS=, read cluster pve_node role node_name ip; do
+    sed 1d nodes.csv | grep -E -v '^#' | grep controlplane | while IFS=, read cluster pve_node role node_name ip; do
         yq -n ".machine.nodeLabels.\"topology.kubernetes.io/region\" = \"$cluster\" | .machine.nodeLabels.\"topology.kubernetes.io/zone\" = \"$node_name\"" \
             > /tmp/labels.yaml
 
@@ -33,7 +31,7 @@ main()
     talosctl config node $CONTROL_PLANE_IP
 
     wait 'Press enter to apply configuration to worker nodes...'
-    sed 1d nodes.csv | grep worker | while IFS=, read cluster pve_node role node_name ip; do
+    sed 1d nodes.csv | grep -E -v '^#' | grep worker | while IFS=, read cluster pve_node role node_name ip; do
         yq -n ".machine.nodeLabels.\"topology.kubernetes.io/region\" = \"$cluster\" | .machine.nodeLabels.\"topology.kubernetes.io/zone\" = \"$node_name\"" \
             > /tmp/labels.yaml
 
@@ -50,7 +48,7 @@ main()
     sleep 30
 
     wait 'Press enter to retrieve the kubeconfig file...'
-    rm "$XDG_CONFIG_HOME/kube/config.$CLUSTER_NAME"
+    [ ! -f "$XDG_CONFIG_HOME/kube/config.$CLUSTER_NAME" ] || rm "$XDG_CONFIG_HOME/kube/config.$CLUSTER_NAME"
     talosctl kubeconfig "$XDG_CONFIG_HOME/kube/config.$CLUSTER_NAME"
     ln -sf "config.$CLUSTER_NAME" "$XDG_CONFIG_HOME/kube/config"
 }
